requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
arrow==1.3.0
async-timeout==4.0.3
asyncer==0.0.8
attrs==25.3.0
Authlib==1.6.3
backoff==2.2.1
backports.asyncio.runner==1.2.0
bidict==0.23.1
binaryornot==0.4.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chainlit==2.7.1.1
chardet==5.2.0
charset-normalizer==3.4.3
chevron==0.14.0
click==8.2.1
colorama==0.4.6
ConfigUpdater==3.2
cookiecutter==2.6.0
cryptography==45.0.7
cuid==0.4
cyclopts==3.23.1
dataclasses-json==0.6.7
Deprecated==1.2.18
dir-tree==0.0.2
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastmcp==2.12.0
fastuuid==0.12.0
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
hf-xet==1.1.8
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.4
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
iniconfig==2.1.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-mcp-adapters==0.1.9
langchain-ollama==0.3.7
langchain-openai==0.3.31
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-checkpoint-sqlite==2.0.11
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.14
Lazify==0.4.0
lazy-object-proxy==1.12.0
litellm==1.76.2
literalai==0.1.201
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.13.1
mcp-python-client==0.1.9
mdurl==0.1.2
monotonic==1.6
more-itertools==10.8.0
multidict==6.6.4
mypy==1.17.1
mypy_extensions==1.1.0
narwhals==2.1.2
nest-asyncio==1.6.0
numpy==2.2.6
ollama==0.5.3
openai==1.99.9
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-instrumentation-alephalpha==0.46.1
opentelemetry-instrumentation-anthropic==0.46.1
opentelemetry-instrumentation-bedrock==0.46.1
opentelemetry-instrumentation-chromadb==0.46.1
opentelemetry-instrumentation-cohere==0.46.1
opentelemetry-instrumentation-crewai==0.46.1
opentelemetry-instrumentation-google-generativeai==0.46.1
opentelemetry-instrumentation-groq==0.46.1
opentelemetry-instrumentation-haystack==0.46.1
opentelemetry-instrumentation-lancedb==0.46.1
opentelemetry-instrumentation-langchain==0.46.1
opentelemetry-instrumentation-llamaindex==0.46.1
opentelemetry-instrumentation-logging==0.57b0
opentelemetry-instrumentation-marqo==0.46.1
opentelemetry-instrumentation-mcp==0.46.1
opentelemetry-instrumentation-milvus==0.46.1
opentelemetry-instrumentation-mistralai==0.46.1
opentelemetry-instrumentation-ollama==0.46.1
opentelemetry-instrumentation-openai==0.46.1
opentelemetry-instrumentation-openai-agents==0.46.1
opentelemetry-instrumentation-pinecone==0.46.1
opentelemetry-instrumentation-qdrant==0.46.1
opentelemetry-instrumentation-redis==0.57b0
opentelemetry-instrumentation-replicate==0.46.1
opentelemetry-instrumentation-requests==0.57b0
opentelemetry-instrumentation-sagemaker==0.46.1
opentelemetry-instrumentation-sqlalchemy==0.57b0
opentelemetry-instrumentation-threading==0.57b0
opentelemetry-instrumentation-together==0.46.1
opentelemetry-instrumentation-transformers==0.46.1
opentelemetry-instrumentation-urllib3==0.57b0
opentelemetry-instrumentation-vertexai==0.46.1
opentelemetry-instrumentation-watsonx==0.46.1
opentelemetry-instrumentation-weaviate==0.46.1
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.57b0
orjson==3.11.2
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.1
parse==1.20.2
pathable==0.4.4
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
pluggy==1.6.0
posthog==3.25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.10.1
pypdf==6.0.0
pyperclip==1.9.0
PyScaffold==4.6
pytest==8.4.1
pytest-asyncio==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.0
rsa==4.9.1
setuptools-scm==9.2.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlite-vec==0.1.6
sse-starlette==3.0.2
starlette==0.47.2
streamlit==1.48.0
syncer==2.0.3
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.2.1
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traceloop-sdk==0.46.1
typer==0.17.3
types-python-dateutil==2.9.0.20250822
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uv==0.8.15
uvicorn==0.35.0
watchdog==6.0.0
watchfiles==0.24.0
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
